+++ src/Game/Text/HelpText.en.sjson.lua
@@ -6,25 +6,25 @@ local order = {
   "DisplayName"
 }
 
--- Add new interaction buttons to the Garden Plots
--- Special interaction/Admire is replaced by Harvest All
+-- Update the Garden's interaction buttons
+-- For planting, Admire is replaced by "Plant All"
+-- For harvesting, "Harvest" is replaced by "Harvest All"
 local newData = {
   {
     Id = "UseGardenPlotPlant",
     DisplayName = "{I} Plant\n {SI} Plant all"
   },
-  -- TODO: When only one plot is ready to be harvested, show Admire again, and change the logic of the called function back to normal admire
   {
     Id = "GardenPlotHarvestAndGift",
-    DisplayName = "{I} Harvest\n {SI} Harvest all\n {G} Water"
+    DisplayName = "{I} Harvest All\n {G} Water"
   },
   {
     Id = "GardenPlotHarvestAndSpecial",
-    DisplayName = "{I} Harvest\n {SI} Harvest All"
+    DisplayName = "{I} Harvest All\n {SI} Admire"
   },
   {
     Id = "GardenPlotHarvestGiftAndSpecial",
-    DisplayName = "{I} Harvest\n {G} Water\n {SI} Harvest All"
+    DisplayName = "{I} Harvest All\n {G} Water\n {SI} Admire"
   }
 }