+++ src/Scripts/InteractPresentation.lua
@@ -1,14 +1,13 @@
 -- Usee is the current plot
 modutil.mod.Path.Wrap("PlantAdmirePresentation", function(base, usee, args)
-	-- If the current plot is ready to harvest or not - admire or continue to harvest/plant all
-	if usee.GrowTimeRemaining > 0 then
+	-- If the current plot has a seed in it, we want to admire it
+	-- Otherwise, we proceed with Plant All logic
+	if usee.SeedName ~= nil then
 		-- Admire
 		base(usee, args)
 		return
-	end
-
-	if usee.SeedName == nil then
-		-- Allows planting multiple seeds at once (disable cooldown)
+	else
+		-- Allows planting multiple seeds at once (disables cooldown)
 		game.SessionState.GlobalCooldowns["UsedGardenPlot"] = nil
 		-- Allows interacting with this plot again (in case the operation was cancelled by closing the inventory without selecting a seed)
 		game.SessionState.GlobalCooldowns[usee.Name .. usee.ObjectId] = nil
@@ -54,25 +53,5 @@ modutil.mod.Path.Wrap("PlantAdmirePresentation", function(base, usee, args)
 			end
 		end
 		RemoveInputBlock({ Name = "GardenOptimizationPlantAllAnimation" })
-	else
-		-- Don't allow any other inputs while harvesting
-		AddInputBlock({ Name = "GardenOptimizationHarvestAllAnimation" })
-		-- The plot is ready to harvest - harvest all
-		-- Harvest the current plot first for presentation purposes
-		game.UseGardenPlot(usee, nil, game.CurrentRun.Hero)
-		-- Allows using the plot again for plantAll right after, and keeps the interact buttons hidden
-		game.SessionState.GlobalCooldowns[usee.Name .. usee.ObjectId] = nil
-		game.wait(0.25)
-
-		-- Harvest all remaining ready plots, skip the others
-		for _, plot in pairs(game.GameState.GardenPlots) do
-			if plot.SeedName ~= nil and plot.GrowTimeRemaining == 0 then
-				-- Reset the cooldown for the function call, as we want to be able to use it multiple times
-				game.SessionState.GlobalCooldowns["UsedGardenPlot"] = nil
-				game.UseGardenPlot(plot, nil, game.CurrentRun.Hero)
-				game.wait(0.25)
-			end
-		end
-		RemoveInputBlock({ Name = "GardenOptimizationHarvestAllAnimation" })
 	end
 end)