+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
-## v0.0.1
+## v1.1.0
 
 <!--Releasenotes start-->
-- Initial release. Adds "Harvest All" and "Plant All" buttons to the Garden.
+- Returned the ability to admire grown plants. "Harvest All" now replaces the "Harvest" instead of the "Admire" prompt.
 <!--Releasenotes end-->
+
+## v1.0.0
+
+- Initial release. Adds "Harvest All" and "Plant All" buttons to the Garden.